Implement local websocket server
This local server would be used as the local signaliing for the initial handshakes (i.e. creating offer and answer) between the RTC peers.
we are running this local ws server on the host machine, becuase we do not want to have a server that will require connection to a remote server via the internet, another option would be to use a QR code to share the SDP between the intending RTC peers, but can't use this since it would require two way scans, hence impacting the users' experience.